backend/requirements.txt
fastapi
uvicorn
sqlmodel
pymysql
python-multipart
python-jose[cryptography]
passlib[bcrypt]
psycopg2-binary
python-dotenv
